update: http-server

+++ http-server/internal/entity/parkingLot.go
@@ -9,7 +9,7 @@ type ParkingLot struct {
 	ID             int     `json:"id"` // Unique identifier for the product
 	Name           string  `json:"name"`
 	BikeCount      int     `json:"bikeCount"`
-	CongestionRate float32 `json:"congestionRate"`
+	CongestionRate float64 `json:"congestionRate"`
 	TotalSpace     int     `json:"totalSpace"`
 	Longitude      float64 `json:"longitude"`
 	Latitude       float64 `json:"latitude"`
